+++ src/audio.rs
@@ -67,15 +67,26 @@ impl Sound {
         })
     }
 
-    /// Creates a new sound from a slice of binary data.
+    /// Creates a new sound from a slice of binary data, encoded in one of Tetra's supported
+    /// file formats.
     ///
     /// This is useful in combination with `include_bytes`, as it allows you to include
     /// your audio data directly in the binary.
     ///
     /// Note that the data is not decoded until playback begins, so this function will not
     /// validate that the data being read is formatted correctly.
-    pub fn from_data(data: &[u8]) -> Sound {
+    pub fn from_file_data(data: &[u8]) -> Sound {
         Sound { data: data.into() }
+    }
+
+    #[deprecated(
+        since = "0.2.13",
+        note = "Renamed to `from_file_data` to disambiguate from other sound data formats (e.g. PCM)."
+    )]
+    #[allow(missing_docs)]
+    #[inline]
+    pub fn from_data(data: &[u8]) -> Sound {
+        Sound::from_file_data(data)
     }
 
     /// Plays the sound.