+++ src/graphics/texture.rs
@@ -3,14 +3,26 @@
 use std::path::Path;
 use std::rc::Rc;
 
-use image::{self, DynamicImage};
+use image;
 
-use crate::error::Result;
+use crate::error::{Result, TetraError};
 use crate::graphics::opengl::{GLTexture, TextureFormat};
 use crate::graphics::{self, DrawParams, Drawable, Rectangle};
 use crate::Context;
 
-/// Texture data.
+/// A 2D texture, held in GPU memory.
+///
+/// The following file formats are supported:
+///
+/// * PNG
+/// * JPEG
+/// * GIF
+/// * BMP
+/// * TIFF
+/// * TGA
+/// * WEBP
+/// * ICO
+/// * PNM
 ///
 /// This type acts as a lightweight handle to the associated graphics hardware data,
 /// and so can be cloned with little overhead.
@@ -33,46 +45,83 @@ impl Texture {
     where
         P: AsRef<Path>,
     {
-        let image = image::open(path)?;
-        Texture::load(ctx, image)
+        let image = image::open(path)?.to_rgba();
+        let (width, height) = image.dimensions();
+
+        Texture::from_rgba(
+            ctx,
+            width as i32,
+            height as i32,
+            image.into_raw().as_slice(),
+        )
     }
 
-    /// Creates a new texture from a slice of binary data.
+    /// Creates a new texture from a slice of data, encoded in one of Tetra's supported
+    /// file formats (except for TGA).
     ///
     /// This is useful in combination with `include_bytes`, as it allows you to include
     /// your textures directly in the binary.
     ///
     /// The format will be determined based on the 'magic bytes' at the beginning of the
     /// data. This should be reasonably reliable, but a `from_data_with_format` function
-    /// might have to be added later.
+    /// might have to be added later. Note that TGA files do not have recognizable magic
+    /// bytes, so this function will not recognize them.
     ///
     /// # Errors
     ///
     /// If the image data was invalid, a `TetraError::Image` will be returned.
-    pub fn from_data(ctx: &mut Context, data: &[u8]) -> Result<Texture> {
-        let image = image::load_from_memory(data)?;
-        Texture::load(ctx, image)
-    }
+    pub fn from_file_data(ctx: &mut Context, data: &[u8]) -> Result<Texture> {
+        let image = image::load_from_memory(data)?.to_rgba();
+        let (width, height) = image.dimensions();
 
-    pub(crate) fn load(ctx: &mut Context, image: DynamicImage) -> Result<Texture> {
-        let rgba_image = image.to_rgba();
-        let (width, height) = rgba_image.dimensions();
-
-        let texture = ctx
-            .gl
-            .new_texture(width as i32, height as i32, TextureFormat::Rgba);
-
-        ctx.gl.set_texture_data(
-            &texture,
-            &rgba_image,
-            0,
-            0,
+        Texture::from_rgba(
+            ctx,
             width as i32,
             height as i32,
-            TextureFormat::Rgba,
-        );
+            image.into_raw().as_slice(),
+        )
+    }
+
+    /// Creates a new texture from a slice of RGBA pixel data.
+    ///
+    /// This is useful if you wish to create a texture at runtime.
+    ///
+    /// Note that this method requires you to provide enough data to fill the texture.
+    /// If you provide too much data, it will be truncated.
+    ///
+    /// # Errors
+    ///
+    /// If not enough data is provided to fill the texture, a `TetraError::NotEnoughData`
+    /// will be returned. This is to prevent OpenGL from reading uninitialized memory.
+    pub fn from_rgba(
+        ctx: &mut Context,
+        width: i32,
+        height: i32,
+        data: &[u8],
+    ) -> Result<Texture> {
+        let expected = (width * height * 4) as usize;
+        let actual = data.len();
+
+        if expected > actual {
+            return Err(TetraError::NotEnoughData { expected, actual });
+        }
+
+        let texture = ctx.gl.new_texture(width, height, TextureFormat::Rgba);
+
+        ctx.gl
+            .set_texture_data(&texture, &data, 0, 0, width, height, TextureFormat::Rgba);
 
         Ok(Texture::from_handle(texture))
+    }
+
+    #[deprecated(
+        since = "0.2.13",
+        note = "Renamed to `from_file_data` to disambiguate from other image data formats (e.g. RGBA)."
+    )]
+    #[allow(missing_docs)]
+    #[inline]
+    pub fn from_data(ctx: &mut Context, data: &[u8]) -> Result<Texture> {
+        Texture::from_file_data(ctx, data)
     }
 
     pub(crate) fn from_handle(handle: GLTexture) -> Texture {